Move compile task integration test
