+++ subprojects/language-cpp/src/test/groovy/dev/nokee/language/cpp/CppSourceSetIntegrationTest.java
@@ -24,7 +24,6 @@ import dev.nokee.language.nativebase.NativeLanguageSourceSetIntegrationTester;
 import dev.nokee.language.nativebase.internal.toolchains.NokeeStandardToolChainsPlugin;
 import dev.nokee.model.internal.ProjectIdentifier;
 import dev.nokee.model.internal.registry.ModelRegistry;
-import org.apache.commons.lang3.StringUtils;
 import org.gradle.api.Project;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
@@ -80,14 +79,6 @@ class CppSourceSetIntegrationTest extends AbstractPluginTest {
 		@Override
 		public String displayName() {
 			return "sources ':zomi'";
-		}
-
-		@Nested
-		class CppCompileTaskTest implements CppCompileTester {
-			@Override
-			public CppCompileTask subject() {
-				return (CppCompileTask) project().getTasks().getByName("compile" + StringUtils.capitalize(variantName()));
-			}
 		}
 	}
 }
