+++ subprojects/language-objective-c/src/test/groovy/dev/nokee/language/objectivec/ObjectiveCCompileTaskIntegrationTest.java
@@ -28,16 +28,24 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static dev.nokee.internal.testing.GradleNamedMatchers.named;
+import static dev.nokee.language.nativebase.internal.NativePlatformFactory.create;
+import static dev.nokee.runtime.nativebase.internal.TargetMachines.of;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 @PluginRequirement.Require(id = "dev.nokee.objective-c-language-base")
-class ObjectiveCCompileTaskIntegrationTest extends AbstractPluginTest {
+class ObjectiveCCompileTaskIntegrationTest extends AbstractPluginTest implements ObjectiveCCompileTester {
 	private ObjectiveCCompileTask subject;
+
+	@Override
+	public ObjectiveCCompileTask subject() {
+		return subject;
+	}
 
 	@BeforeEach
 	void createSubject() {
 		project.getPluginManager().apply(NokeeStandardToolChainsPlugin.class);
 		subject = project.getExtensions().getByType(ModelRegistry.class).register(project.getExtensions().getByType(ObjectiveCSourceSetRegistrationFactory.class).create(LanguageSourceSetIdentifier.of(ProjectIdentifier.of(project), "gugo"))).element("compile", ObjectiveCCompileTask.class).get();
+		subject.getTargetPlatform().set(create(of("macos-x64")));
 	}
 
 	@Test
